Moved gps rtos task to separate file and addressed tabs to space comment
