+++ firmware/Core/Inc/rtos_tasks/gps_rtos_tasks.h
@@ -1,0 +1,9 @@
+
+#ifndef __INCLUDE_GUARD__GPS_RTOS_TASKS_H__
+#define __INCLUDE_GUARD__GPS_RTOS_TASKS_H__
+
+#include "main.h"
+
+void TASK_receive_gps_info(void *argument);
+
+#endif // __INCLUDE_GUARD__GPS_RTOS_TASKS_H__
